修复审查发现的高危缺陷：Redis 读路径 fail-open、伪造 jti 致命错误、刷新令牌越权
安全
- RedisTokenStorage::isBlacklisted() 用 (bool) exists() 判断，phpredis 在超时/链路错误时
  返回 false，与"0 个键"无法区分 → 已拉黑令牌被放行，且不抛异常就绕过了 fail_open 的
  fail-closed 策略。改为 false 即抛 STORAGE_ERROR、> 0 才算拉黑
- 刷新令牌可当访问令牌用：decode() 不校验 token_type，Bearer <refresh token> 能直接访问
  受保护接口，而刷新令牌有效期更长。decode() / validate() 默认拒绝 token_type=refresh，
  需要时显式传 $allowRefresh = true；refresh() / blacklist() 内部已放开。
  注意：这是行为变更，若你有意用 decode() 读取刷新令牌，需补第二个参数
- JWT::isBlacklisted() 对未验签 payload 里的 {"jti":[...]} 会抛 TypeError（Error 接不住）
  → 未认证请求可以打成 500。解析时丢弃非字符串 jti，捕获放宽到 \Throwable

缺陷
- 非十六进制 jti（如迁移过来的 UUID）在 file/redis/memcached 抛 "Invalid JTI format"：
  fail_open=false 时合法令牌全被 401，true 时该令牌的黑名单永不生效，database 却正常。
  改为统一归一化（file 用 bin2hex、redis/memcached 用 sha256），已有十六进制 jti 的键名
  不变，路径穿越同时消除
- firebase v7 允许无 exp 令牌，blacklist() / refresh() 直接把 $payload['exp'] 当 int 用会
  抛 TypeError，refresh() 还会在换发新令牌前中断。改为 (int) ($payload['exp'] ?? 0)，
  且仅在 exp > time() 时写黑名单
- memcached 默认合并里的 'servers' => [] 让 ?? 默认值永不生效 → addServers([]) 后
  Laravel/webman/原生 的 memcached 存储完全不可用。改用 ?:
- Laravel ServiceProvider 与 webman Middleware 在数组字面量里立即取 PDO，用 file/redis
  存储的应用也会因为没装 webman/database 或数据库故障而整体不可用。改为与 ThinkPHP/
  Hyperf 一致，按 storage.type 取连接
- auto_cleanup 的 static $lastCleanup 在 PHP-FPM 下每请求重置，cleanup_interval 形同虚设
  （每请求清理一次）。改用时间戳文件跨请求节流，抽出 cleanupDue() 并可单测
- 算法未校验：JWT_ALGORITHM=hs256 拼错能构造成功，之后 encode() 抛 DomainException、
  decode() 把全部合法令牌判为无效。构造时按 FirebaseJWT::$supported_algs 校验

配置与文档
- 删除无人读取的 storage.database（README 曾写作"Redis 数据库编号"）
- 统一 webman / 原生配置模板的自动清理环境变量名（原 JWT_ADVANCED_*，与其余三个框架
  和文档不一致）；README 补上 servers / options
- 删除与 Laravel 逐字节相同且在 autoload.files 中永远被 function_exists 挡掉的
  ThinkPHP/helpers.php
- bearerToken() 容忍多余空白；writeEnvSecret() 返回成功与否，安装命令在 .env 缺失时
  明确提示而不是打印成功；转义替换串中的反斜杠
- FileTokenStorage::cleanup() 回收崩溃残留的 *.json.tmp.*；gc_probability < 0.01 由
  "永不触发"改为按概率触发；expire_time 由 INT 改 BIGINT
- README 新增「刷新令牌不能当访问令牌用」「jti 格式」两节，Hyperf 补上下文读 payload，
  修正 auto_cleanup 说明

测试：229 用例 / 433 断言通过（9 个 memcached 用例因扩展缺失跳过），
      新增刷新令牌拒绝、伪造 jti、算法校验、Redis fail-open、jti 归一化、清理节流等用例

+++ src/erik-jwt/DatabaseTokenStorage.php
@@ -42,7 +42,7 @@ class DatabaseTokenStorage implements TokenStorageInterface
     {
         $sql = "CREATE TABLE IF NOT EXISTS {$this->tableName} (
             jti VARCHAR(64) PRIMARY KEY,
-            expire_time INT NOT NULL,
+            expire_time BIGINT NOT NULL,
             created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP
         )";
